+++ packages/ts-newskit/src/components/puzzles/offers-banner/__tests__/OffersBanner.test.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { OffersBanner } from '../index';
+import { render } from '../../../../utils/test-utils';
+
+const renderComponent = () => render(<OffersBanner />);
+
+describe('OffersBanner', () => {
+  it('should render OffersBanner component', () => {
+    const { asFragment } = renderComponent();
+    expect(asFragment()).toMatchSnapshot();
+  });
+});