+++ packages/ts-newskit/src/components/slices/lead-article/index.tsx
@@ -54,7 +54,7 @@ export interface LeadArticleProps {
   imageTop?: boolean;
   hasTopBorder?: boolean;
   contentTop?: boolean;
-  contentWidth?: string;
+  contentWidth?: MQ<string> | string;
   headlineTypographyPreset?: MQ<string> | string;
   loadingAspectRatio?: string;
   imageMarginBlockStart?: string;
@@ -104,7 +104,7 @@ export const LeadArticle = ({
     imageWithCorrectRatio &&
     imageWithCorrectRatio.url !== '' && {
       media: {
-        src: imageWithCorrectRatio.url,
+        src: `${imageWithCorrectRatio.url}&resize=750`,
         alt: (images && images.alt) || headline,
         loadingAspectRatio: imageWithCorrectRatio.ratio
       }
@@ -148,9 +148,7 @@ export const LeadArticle = ({
           : `content media`
       }}
       columnGap="space040"
-      columns={{
-        md: displayArticleVertical ? '100%' : `${contentWidth || '260px'} auto`
-      }}
+      columns={displayArticleVertical || !contentWidth ? '100%' : contentWidth}
     >
       {hasImage &&
         !hideImage && (