+++ packages/ts-newskit/src/slices/lead-story-3/index.tsx
@@ -78,13 +78,11 @@ export const LeadStory3 = ({
           </FullWidthBlock>
         </Hidden>
         <Block marginBlockEnd={{ xs: 'space040', md: 'space000' }}>
-          <Visible lg xl>
-            <LeadStoryDivider
-              overrides={{ stylePreset: 'lightDivider' }}
-              vertical
-              position="right"
-            />
-          </Visible>
+          <LeadStoryDivider
+            overrides={{ stylePreset: 'lightDivider' }}
+            vertical
+            position="right"
+          />
           <LeadArticle
             article={modifiedLeadArticle}
             clickHandler={clickHandler}