+++ packages/user-state/src/states.js
@@ -1,8 +1,6 @@
 export const SUBSCRIBER = "Subscriber";
-export const RA = "Registered Access";
-export const RA_EXPIRED = "Registered Access (Expired)";
 export const GUEST = "Guest";
 
-const USER_STATES = { SUBSCRIBER, RA, RA_EXPIRED, GUEST };
+const USER_STATES = { SUBSCRIBER, GUEST };
 
 export default USER_STATES;