+++ packages/article-magazine-standard/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.72.67](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.66...@times-components/article-magazine-standard@3.72.67) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.66](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.65...@times-components/article-magazine-standard@3.72.66) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.65](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.64...@times-components/article-magazine-standard@3.72.65) (2023-10-17)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.64](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.63...@times-components/article-magazine-standard@3.72.64) (2023-10-17)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
+## [3.72.63](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.62...@times-components/article-magazine-standard@3.72.63) (2023-10-16)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
 ## [3.72.62](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@3.72.61...@times-components/article-magazine-standard@3.72.62) (2023-10-16)
 
 **Note:** Version bump only for package @times-components/article-magazine-standard