+++ packages/interactive-wrapper/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.45](https://github.com/newsuk/times-components/compare/@times-components/interactive-wrapper@0.9.44...@times-components/interactive-wrapper@0.9.45) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/interactive-wrapper
+
+
+
+
+
+## [0.9.44](https://github.com/newsuk/times-components/compare/@times-components/interactive-wrapper@0.9.43...@times-components/interactive-wrapper@0.9.44) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/interactive-wrapper
+
+
+
+
+
+## [0.9.43](https://github.com/newsuk/times-components/compare/@times-components/interactive-wrapper@0.9.42...@times-components/interactive-wrapper@0.9.43) (2023-10-17)
+
+**Note:** Version bump only for package @times-components/interactive-wrapper
+
+
+
+
+
 ## [0.9.42](https://github.com/newsuk/times-components/compare/@times-components/interactive-wrapper@0.9.41...@times-components/interactive-wrapper@0.9.42) (2023-08-18)
 
 **Note:** Version bump only for package @times-components/interactive-wrapper