+++ packages/key-facts/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.12.37](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.12.36...@times-components/key-facts@2.12.37) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/key-facts
+
+
+
+
+
+## [2.12.36](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.12.35...@times-components/key-facts@2.12.36) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/key-facts
+
+
+
+
+
+## [2.12.35](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.12.34...@times-components/key-facts@2.12.35) (2023-10-17)
+
+**Note:** Version bump only for package @times-components/key-facts
+
+
+
+
+
 ## [2.12.34](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.12.33...@times-components/key-facts@2.12.34) (2023-10-13)
 
 **Note:** Version bump only for package @times-components/key-facts