+++ packages/storybook/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.11.2](https://github.com/newsuk/times-components/compare/@times-components/storybook@4.11.1...@times-components/storybook@4.11.2) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/storybook
+
+
+
+
+
+## [4.11.1](https://github.com/newsuk/times-components/compare/@times-components/storybook@4.11.0...@times-components/storybook@4.11.1) (2023-10-17)
+
+**Note:** Version bump only for package @times-components/storybook
+
+
+
+
+
 # [4.11.0](https://github.com/newsuk/times-components/compare/@times-components/storybook@4.10.6...@times-components/storybook@4.11.0) (2023-08-18)