+++ packages/topic/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.10.37](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.36...@times-components/topic@5.10.37) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [5.10.36](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.35...@times-components/topic@5.10.36) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [5.10.35](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.34...@times-components/topic@5.10.35) (2023-10-17)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
 ## [5.10.34](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.33...@times-components/topic@5.10.34) (2023-10-13)
 
 **Note:** Version bump only for package @times-components/topic