+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.16.20](https://github.com/newsuk/times-components/compare/@times-components/utils@6.16.19...@times-components/utils@6.16.20) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/utils
+
+
+
+
+
+## [6.16.19](https://github.com/newsuk/times-components/compare/@times-components/utils@6.16.18...@times-components/utils@6.16.19) (2023-10-18)
+
+**Note:** Version bump only for package @times-components/utils
+
+
+
+
+
+## [6.16.18](https://github.com/newsuk/times-components/compare/@times-components/utils@6.16.17...@times-components/utils@6.16.18) (2023-10-17)
+
+**Note:** Version bump only for package @times-components/utils
+
+
+
+
+
 ## [6.16.17](https://github.com/newsuk/times-components/compare/@times-components/utils@6.16.16...@times-components/utils@6.16.17) (2023-08-18)
 
 **Note:** Version bump only for package @times-components/utils